Move edit line to correct folder
