+++ app/dashboard/lines/[id]/edit/ui/InformationDialog.tsx
