+++ app/dashboard/lines/[id]/edit/ui/ShowInfoButton.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import {useState} from "react";
-import InformationDialog from "@/app/dashboard/lines/test-page/InformationDialog";
+import InformationDialog from "@/app/dashboard/lines/[id]/edit/ui/InformationDialog";
 
 export default function ShowInfoButton() {
