+++ app/lib/definitions.ts
@@ -1,11 +1,3 @@
-import { Position } from "../dashboard/lines/test-page/lib/new-definitions";
-
-export type Line = {
-    id: string;
-    created_at: string;
-    name: string;
-}
-
 export type Agencies = {
     id: number;
     created_at: string;
@@ -20,11 +12,6 @@ export interface Stop {
     position: Position
 }
 
-export interface Location {
-    type: string;
-    coordinates: number[]
-}
-
 // Map the Agency type to the Option type using a mapped type
 type AgencyToOptionMap = {
     [K in keyof Agencies]: K extends 'id' ? 'value' : K extends 'name' ? 'label' : never;
@@ -37,32 +24,28 @@ export interface UserInfo {
     role: string
 }
 
-//NEW DEFINITIONS FOR MAPS AND DB
-
-export interface Point {
-    lat: number,
-    lng: number
+export interface Position {
+    lat: number;
+    lng: number;
 }
 
-export interface Polyline {
-    points: Point[]
-}
-
-export interface NewStop {
-    id: number,
-    location: Point
-}
-
-export interface Step {
-    id: number,
-    sequence_number: number,
-    polyline: Polyline,
-    startStop: NewStop,
-    endStop: NewStop
+export interface RoutePoint {
+    id: number | null;
+    position: Position;
+    isStop: boolean;
+    order: number;
+    busStop: Stop | null;
 }
 
 export interface Route {
-    id: number,
-    name: string,
-    steps: Step[]
+    id: number;
+    created_at: string;
+    updated_at: string;
+    line_number: string;
+    legacy_line_number: string;
+    units: number;
+    agency_id: number;
+    transport_type: string;
+    line_type: string;
+    points: RoutePoint[] | [];
 }
